Add a requeue channel
As a controller processes events, it may not be able to act successfully
on these events: there may be a transient problem when writing to the
API server, or the system may not be in a converged state and it may be
necessary to reprocess the update at a later time.

The `requeue` channel provides a bounded, multi-producer/single-
consumer, delay-queue channel. Applications can use this channel to
reschedule updates for a later time. While this is especially useful for
controllers, this channel is not inherently tied to Kubernetes in any
way.