shutdown: Bias towards processing remaining work (#15)
`shutdown::CancelOnShutdown` wraps a `Future` or `Stream` so that the
task completes when shutdown is signaled. Currently, this is done by
stopping all additional work on these tasks as soon as the shutdown
signal is observed. This makes it difficult to ensure that in-flight
work is processed during shutdown.

This change alters the behavior so that `CancelOnShutdown` polls the
inner `Future` or `Stream` before polling the shutdown watch, so that
the shutdown watch is only polled when the inner instance is pending.
This will, in-turn, block the shutdown and runtime futures from
completing until all pending work is processed.

In pathological cases, this could prevent the process from terminating
gracefully. It seems better to the user or kubelet abort the application
if it doesn't complete its work in a timely fashion.