Change `index` to keep track of its keys
Rather than force the indexing type to keep track of its own key
structures, it's better for the indexer to do this internaly. This makes
it so that the hash implementation isn't a part of the public interface;
and also makes it so that the index implementation need not track all
resources for proper operations.

This is a breaing change that removes the `snapshot_keys` method from
the `IndexClusterResource` and `IndexNamespacedResource` traits. As
such, the crate version is bumped to v0.4.0