index: Allow indexers to handle restarts (#28)
Indexers may want to handle batch updates specially, for instance to
run an expensive computation after the batch update is applied. With the
current `Index` trait APIs, this is not possible.

This change adds `IndexClusterResources::reset` and
`IndexNamespacedResources::reset` methods with a default implementation
that preserves the current behavior: `apply` is called for each live
resource and `delete` is called for all resources that aren't included
in the update. This default trait-method implementation ensures this
change is backwards-compatible.