docs: fix `index` module not being present on docs.rs
Currently, the docs.rs documentation for `kubert` doesn't show the
`kubert::index` module, or any of its contents (such as the
`Index{Cluster, Namespaced}Resource` traits).

This is because the crate's docs.rs metadata currently enables a
specific list of features, which doesn't include the `index` feature.
Presumably, the list of features to enable when building docs was not
modified when the `index` module was added.

This commit fixes this by changing the docs.rs metadata so that
`--all-features` is passed instead of a list of feature flags. This way,
not only is the `index` module now present, but also, if any new feature
flags are added, the docs.rs metadata won't need to be modified again.

We still pass a list of explicit feature flags to enable
`k8s-openapi/v1_25`, since `--all-features` only effects *`kubert`'s*
features, and not those of dependencies.