lease: A distributed, cooperative lock
Applications that manage state in Kubernetes--for instance, those that
update resource statuses, may need to coordinate access to that state so
that only one replica is trying to update resources at a time.

The module manages coordination.k8s.io/v1 Lease resources to ensure that
only a single claimant owns the lease.