lease: don't terminate `LeaseManager` on `ensure_claimed` errors
Currently, if the `ensure_claimed` call in the `LeaseManager::spawn`
background task fails, the entire task will bail and terminate
immediately. This is not desirable, as the intended use of
`LeaseManager::spawn` is to spawn a single background task that will
manage leasses for the entire lifetime of a controller process, and it
should not terminate in the face of a transient error claiming the
lease.

This branch changes `LeaseManager::spawn` to log the error and continue
looping, rather than bailing out. We may want to consider adding some
kind of retry backoff logic here as well, but this solution still seems
better than simply bailing o34ut.

Fixes #134