errors: add `ExponentialBackoff` implementation
This branch adds an implementation of exponential jittered backoffs in
the `errors` module. This implementation is taken from the
`linkerd-exp-backoff` crate in Linkerd. Exponential backoffs will be
used in PR #136 when retrying transient errors that occur when
acquiring/renewing a lease.

I thought it might be desirable to make this a new public API, so that
controllers implemented using `kubert` can add exponential backoffs for
other operations without duplicating code that exists in `kubert`
internally. If this isn't an API we want `kubert` to provide, though,
I'm happy to make this module private. An alternative solution would be
to publish the `linkerd-exp-backoff` crate as a standalone crate that
`kubert` could depend on from Cargo, but I thought we would probably
prefer not to have to manage publishing a new small standalone crate
from the `linkerd2-proxy` repo (which isn't really set up to publish
crates at the moment). However, if we'd rather just do that, I can also
set up that crate to be publishable.