server: add test for `load_tls`
This commit adds a test for the `tls::load_tls` function in the `server`
module. This test passes with `--features rustls-tls` but fails with
`--features boring-tls`, reproducing a bug where the BoringSSL version
of `load_tls` passes the entire PEM file to
`boring::PKey::private_key_from_pkcs8`, which expects a single PKCS8
key, and not the entire PEM file.