chore: mute dead code warnings on handle permit
the `Handle` type wraps an inner semaphore permit. it is never read,
because simply _holding_ it is itself semantically meaningful.

this does, however, cause a warning when compiled in Rust 1.80:

```text
error: field `0` is never read
  --> kubert/src/initialized.rs:25:19
   |
25 | pub struct Handle(OwnedSemaphorePermit);
   |            ------ ^^^^^^^^^^^^^^^^^^^^
   |            |
   |            field in this struct
   |
   = note: `Handle` has a derived impl for the trait `Debug`, but this is intentionally ignored during dead code analysis
```

this commit introduces a `#[allow(dead_code)]` directive on that inner
field, to address this warning and indicate explicitly that this field
is never read.