🦀 chore: fix build errors for 1.80 toolchain (#263)
* chore: `lints.rust` allows `cfg(tokio_unstable)`

building `kubert` with Rust 1.80 or later will fail, with an error
shaped like this:

```text
error: unexpected `cfg` condition name: `tokio_unstable`
  --> kubert-prometheus-tokio/src/lib.rs:13:27
   |
13 | #[cfg(all(feature = "rt", tokio_unstable))]
   |                           ^^^^^^^^^^^^^^
   |
   = help: consider using a Cargo feature instead
   = help: or consider adding in `Cargo.toml` the `check-cfg` lint config for the lint:
            [lints.rust]
            unexpected_cfgs = { level = "warn", check-cfg = ['cfg(tokio_unstable)'] }
   = help: or consider adding `println!("cargo::rustc-check-cfg=cfg(tokio_unstable)");` to the top of the `build.rs`
   = note: see <https://doc.rust-lang.org/nightly/rustc/check-cfg/cargo-specifics.html> for more information about checking conditional configuration

error: could not compile `kubert-prometheus-tokio` (lib) due to 3 previous errors
```

see the announcement here, from the Cargo team, as of
nightly-2024-05-05:
<https://blog.rust-lang.org/2024/05/06/check-cfg.html#expecting-custom-cfgs>

this commit adds a `[lints.rust]` entry adding the `tokio_unstable` flag
to the list of expected cfg's.

* chore: mute dead code warnings on handle permit

the `Handle` type wraps an inner semaphore permit. it is never read,
because simply _holding_ it is itself semantically meaningful.

this does, however, cause a warning when compiled in Rust 1.80:

```text
error: field `0` is never read
  --> kubert/src/initialized.rs:25:19
   |
25 | pub struct Handle(OwnedSemaphorePermit);
   |            ------ ^^^^^^^^^^^^^^^^^^^^
   |            |
   |            field in this struct
   |
   = note: `Handle` has a derived impl for the trait `Debug`, but this is intentionally ignored during dead code analysis
```

this commit introduces a `#[allow(dead_code)]` directive on that inner
field, to address this warning and indicate explicitly that this field
is never read.