feat(runtime): add watch admin diagnostics
The runtime is used to instrument watches with metrics; but these metrics are
necessarily low-granularity. When debugging the state of a controller, it can be
useful to get a higher-fidelity snapshot of the runtime's state.

By inspecting the states of all watches, we can surface valuable diagnostic
information without requiring applications to instrument this themselves.

This change adds a new `runtime-diagnostics` feature. When it is enabled,
watches are instrumented so that their states can be queried at runtime via the
/kubert.json admin endpoint. This endpoint is only is served to requests that
originate on the loopback interface to prevent this state from being exported to
arbitrary clients.