chore(client,server)!: remove openssl support
We have some fairly complicated feature flagging to support openssl in addition
to rustls.

Now that rustls itself is a facade that can be used with a variety of
cryptographic providers (including OpenSSL), this change removes the
rustls-tls and openssl-tls features.

Rustls support is always enabled with the client and server features.