feat(admin): add optional diagnostics endpoint
This change adds a new /kubert.json endpoint on the admin HTTP server. This
endpoint returns a JSON object with various diagnostic information about the
kubert runtime that do not readily fit into metrics. This admin endpoint is only
available from the localhost and cannot be accessed from the non-loopback
network.

The initial version of this endpoint includes only timestamp information.
Additional diagnostic information will be added in future commits.