feat(runtime): add watch admin diagnostics (#317)
The runtime is used to instrument watches with metrics; but these metrics are
necessarily low-granularity. When debugging the state of a controller, it can be
useful to get a higher-fidelity snapshot of the runtime's state.

By inspecting the states of all watches, we can surface valuable diagnostic
information without requiring applications to instrument this themselves.

This change updates the runtime-diagnostics feature so that, when enabled,
watch states are automatically tracked and included in the diagnostics endpoint.