feat(client): add a default response headers timeout (#347)
The underlying kube client does not enforce any request timeouts, relying on the
server to implement them. But this means that calls--especially those dispatched
by a watcher--may hang indefinitely waiting for an API server response.

This change introduces a default timeout for response headers, which is set to
9s. Furthermore, the ClientArgs type is used to expose this timeout
configuration as a command-line option, kube-api-response-headers-timeout.